+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## unreleased
+### Fixed
+- Fix parameters being validated in the wrong order
+
 ## [0.4.0] - 2016-12-30
 ### Fixed
 - Fixed basic permissions being incorrectly allowed, e.g. `/articles:read,update` would "allow" `/articles:all`.