+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 - Refactored test files setup
+- Refactor `export default permission` to `export { permission }`
 
 ## [0.4.0] - 2016-12-30
 ### Fixed