+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - Refactored test files setup
 - Refactor `export default permission` to `export { permission }`
 - Changed "slow" test indicator from default to 10ms
+- `permission()` constructor now accepts an URLPermission object and clones it
 
 ### Added
 - `permissions()` function and companion class `URLPermissions` which evaluates a collection of permissions